Moving revoke tests into their own file
